backend/requirements.txt
Werkzeug==0.15.2
mysqlclient==1.3.13
Django==2.2
django-mysql==3.0.0.post1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.0.0
django-cors-headers==2.4.0
boto3==1.9.102
requests==2.21.0
PyJWT==1.7.1
watchtower==0.5.5
cryptography==2.6.1
requests==2.21.0
